+++ lib/src/components/data-table/DataTable.Body.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 
 import { Table } from '../table'
-import { DataTable } from './'
+import { DataTable } from '.'
 import { useDataTable } from './DataTableContext'
 type DataTableBodyProps = Omit<
   React.ComponentProps<typeof Table.Body>,
